+++ ragondin/components/indexer/indexer.py
@@ -1,3 +1,4 @@
+import asyncio
 import gc
 import inspect
 from typing import Dict, List, Optional
@@ -6,12 +7,11 @@ import ray
 import torch
 from config import load_config
 from langchain_core.documents.base import Document
-from ray.util.state import get_task
+from langchain_openai import OpenAIEmbeddings
+from loguru import logger
+from ray.util.actor_pool import ActorPool
 
-from ..reranker import Reranker
-from ..utils import SingletonMeta
 from .chunker import ABCChunker, ChunkerFactory
-from .embeddings import HFEmbedder
 from .loaders.loader import DocSerializer
 from .vectordb import ConnectorFactory
 
@@ -20,44 +20,57 @@ config = load_config()
 
 # Set ray resources
 NUM_GPUS = config.ray.get("num_gpus")
-NUM_CPUS = config.ray.get("num_cpus")
-N_PARALLEL_INDEXATION = config.ray.get("n_parallel_indexation")
+POOL_SIZE = config.ray.get("pool_size")
+MAX_TASKS_PER_WORKER = config.ray.get("max_tasks_per_worker")
 
 if torch.cuda.is_available():
-    gpu, cpu = NUM_GPUS, NUM_CPUS
+    gpu = NUM_GPUS
 else:
-    gpu, cpu = 0, NUM_CPUS
+    gpu = 0
 
 
-@ray.remote(
-    num_cpus=cpu,
-    num_gpus=gpu,
-    max_task_retries=2,
-    max_restarts=-1,
-    concurrency_groups={"compute": N_PARALLEL_INDEXATION},
-)
-class Indexer(metaclass=SingletonMeta):
-    def __init__(self, config, logger, device=None):
-        self.config = config
-        self.enable_insertion = self.config.vectordb["enable"]
-        self.embedder = HFEmbedder(embedder_config=config.embedder, device=device)
-        self.serializer = DocSerializer(data_dir=config.paths.data_dir, config=config)
+@ray.remote(num_gpus=gpu, max_task_retries=2)
+class IndexerWorker:
+    """This class bridges static files with the vector store database.*"""
+
+    def __init__(self) -> None:
+        """
+        Initializes the Indexer class with the given configuration, logger, and optional device.
+
+        Args:
+            config (Config): Configuration object containing settings for the embedder, paths, llm, and insertion.
+            logger (Logger): Logger object for logging information.
+            device (str, optional): Device to be used by the embedder. Defaults to None.
+        """
+        from config import load_config
+        from langchain_openai import OpenAIEmbeddings
+        from loguru import logger
+
+        from .vectordb import ConnectorFactory
+
+        self.config = load_config()
+        self.embedder = OpenAIEmbeddings(
+            model=self.config.embedder.get("model_name"),
+            base_url=self.config.embedder.get("base_url"),
+            api_key=self.config.embedder.get("api_key"),
+        )
+        self.serializer = DocSerializer(
+            data_dir=self.config.paths.data_dir, config=self.config
+        )
         self.chunker: ABCChunker = ChunkerFactory.create_chunker(
-            config, embedder=self.embedder.get_embeddings()
+            self.config, embedder=self.embedder
         )
         self.vectordb = ConnectorFactory.create_vdb(
-            config, logger=logger, embeddings=self.embedder.get_embeddings()
+            self.config, logger=logger, embeddings=self.embedder
         )
         self.logger = logger
-
-        # reranker
-        self.reranker = None
-        self.reranker_enabled = config.reranker["enable"]
-        if self.reranker_enabled:
-            self.logger.debug(f"Reranker enabled: {self.reranker_enabled}")
-            self.reranker = Reranker(self.logger, config)
-
-        self.logger.info("Indexer initialized...")
+        self.default_partition = "_default"
+        self.enable_insertion = self.config.vectordb["enable"]
+        self.task_state_manager = ray.get_actor(
+            "TaskStateManager", namespace="ragondin"
+        )
+        self.task_count = 0
+        self.logger.info("Indexer worker initialized.")
 
     async def serialize(self, path: str, metadata: Optional[Dict] = {}):
         self.logger.info(f"Starting serialization of documents from {path}...")
@@ -77,20 +90,30 @@ class Indexer(metaclass=SingletonMeta):
             self.logger.info(f"No chunks for {file_path}")
             return []
 
-    @ray.method(concurrency_group="compute")
     async def add_file(
         self,
+        task_id: str,
         path: str | list[str],
         metadata: Optional[Dict] = {},
         partition: Optional[str] = None,
     ):
+        self.task_count += 1
+
         partition = self._check_partition_str(partition)
         metadata = {**metadata, "partition": partition}
+
+        # Serialize document
+        self.task_state_manager.set_state.remote(task_id, "SERIALIZING")
         doc = await self.serialize(path, metadata=metadata)
+
+        # Chunk docs
+        self.task_state_manager.set_state.remote(task_id, "CHUNKING")
         chunks = await self.chunk(doc, path)
 
+        # Add chunks to the vector database
         try:
             if self.enable_insertion:
+                self.task_state_manager.set_state.remote(task_id, "INSERTING")
                 await self.vectordb.async_add_documents(chunks)
                 self.logger.debug(f"Documents {path} added.")
             else:
@@ -106,12 +129,84 @@ class Indexer(metaclass=SingletonMeta):
                 torch.cuda.empty_cache()
                 torch.cuda.ipc_collect()
 
+    def _check_partition_str(self, partition: Optional[str]):
+        if partition is None:
+            self.logger.warning("Partition not provided. Using default partition.")
+            partition = self.default_partition
+        elif not isinstance(partition, str):
+            raise ValueError("Partition should be a string.")
+        return partition
+
+    async def get_task_count(self):
+        return self.task_count
+
+
+@ray.remote(max_restarts=-1)
+class Indexer:
+    def __init__(self):
+        config = load_config()
+        self.config = config
+        self.logger = logger
+        self.enable_insertion = self.config.vectordb["enable"]
+        self.embedder = OpenAIEmbeddings(
+            model=self.config.embedder.get("model_name"),
+            base_url=self.config.embedder.get("base_url"),
+            api_key=self.config.embedder.get("api_key"),
+        )
+        self.vectordb = ConnectorFactory.create_vdb(
+            config, logger, embeddings=self.embedder
+        )
+        self.task_state_manager = ray.get_actor(
+            "TaskStateManager", namespace="ragondin"
+        )
+        self.queue = ray.get_actor("IndexerQueue", namespace="ragondin")
+        logger.info("Indexer supervisor actor initialized.")
+
+    async def add_file(self, path, metadata, partition):
+        # Retrieve task_id from the Ray runtime context
+        task_id = ray.get_runtime_context().get_task_id()
+
+        self.task_state_manager.set_state.remote(task_id, "QUEUED")
+
+        # Get a free worker from the pool
+        worker = await self.queue.get_worker.remote()
+
+        try:
+            # Send task to the worker
+            await worker.add_file.remote(task_id, path, metadata, partition)
+
+        except Exception as e:
+            self.logger.error(f"Error in `add_file` for path {path}: {e}")
+            self.task_state_manager.set_state.remote(task_id, "FAILED")
+            raise
+        finally:
+            task_count = await worker.get_task_count.remote()
+            if task_count >= MAX_TASKS_PER_WORKER:
+                logger.info("Worker reached max task count, restarting...")
+                ray.kill(worker)
+                worker = IndexerWorker.remote()
+            self.queue.push.remote(worker)
+        self.task_state_manager.set_state.remote(task_id, "COMPLETED")
+
     def delete_file(self, file_id: str, partition: str):
+        """
+        Deletes files from the vector database based on the provided filters.
+        Args:
+            filters (Union[Dict, List[Dict]]): A dictionary or list of dictionaries containing the filters to identify files to be deleted.
+            collection_name (Optional[str]): The name of the collection from which files should be deleted. Defaults to None.
+        Returns:
+            Tuple[List[Dict], List[Dict]]: A tuple containing two lists:
+                - deleted_files: A list of filters for the files that were successfully deleted.
+                - not_found_files: A list of filters for the files that were not found in the database.
+        Raises:
+            Exception: If an error occurs during the deletion process, it is logged and the function continues with the next filter.
+        """
         if not self.enable_insertion:
             self.logger.error(
                 "Vector database is not enabled, however, the delete_files method was called."
             )
             return
+
         try:
             # Get points associated with the file name
             points = self.vectordb.get_file_points(file_id, partition)
@@ -169,7 +264,7 @@ class Indexer(metaclass=SingletonMeta):
         self,
         query: str,
         top_k: int = 5,
-        similarity_threshold: int = 0.6,
+        similarity_threshold: int = 0.80,
         partition: Optional[str | List[str]] = None,
         filter: Optional[Dict] = {},
     ) -> List[Document]:
@@ -182,22 +277,14 @@ class Indexer(metaclass=SingletonMeta):
             similarity_threshold=similarity_threshold,
             filter=filter,
         )
-
-        # if self.reranker_enabled:
-        #     self.logger.debug(f"Reranker enabled: {self.reranker_enabled}")
-        #     results = await self.reranker.rerank(query, documents=results, top_k=top_k)
-
         return results
-
-    def check_file_exists_in_partition(self, file_id: str, partition: str):
-        return self.vectordb.file_exists(file_id=file_id, partition=partition)
 
     def delete_partition(self, partition: str):
         return self.vectordb.delete_partition(partition)
 
-    def get_task_status(self, task_id: str):
+    async def get_task_status(self, task_id: str):
         """Get the status of a task."""
-        return get_task(task_id)
+        return await self.task_state_manager.get_state.remote(task_id)
 
     def _check_partition_list(self, partition: Optional[str]):
         if partition is None:
@@ -209,14 +296,6 @@ class Indexer(metaclass=SingletonMeta):
             not all(isinstance(p, str) for p in partition)
         ):
             raise ValueError("Partition should be a string or a list of strings.")
-        return partition
-
-    def _check_partition_str(self, partition: Optional[str]):
-        if partition is None:
-            self.logger.warning("Partition not provided. Using default partition.")
-            partition = self.default_partition
-        elif not isinstance(partition, str):
-            raise ValueError("Partition should be a string.")
         return partition
 
     async def _delegate_vdb_call(self, method_name: str, *args, **kwargs):
@@ -234,3 +313,42 @@ class Indexer(metaclass=SingletonMeta):
         """Helper method to check if a vectordb method is async."""
         method = getattr(self.vectordb, method_name, None)
         return inspect.iscoroutinefunction(method) if method else False
+
+
+@ray.remote
+class TaskStateManager:
+    def __init__(self):
+        self.states = {}
+        self.lock = asyncio.Lock()
+
+    async def set_state(self, task_id: str, state: str):
+        async with self.lock:
+            self.states[task_id] = state
+
+    async def get_state(self, task_id: str) -> Optional[str]:
+        async with self.lock:
+            state = self.states.get(task_id, None)
+            return state
+
+    async def get_all_states(self):
+        async with self.lock:
+            return dict(self.states)
+
+
+@ray.remote(concurrency_groups={"worker_queue": 1})
+class IndexerQueue:
+    def __init__(self):
+        self.pool = ActorPool([IndexerWorker.remote() for _ in range(POOL_SIZE)])
+
+    @ray.method(concurrency_group="worker_queue")
+    async def get_worker(self):
+        """Get a free worker from the pool."""
+        while True:
+            if self.pool.has_free():
+                worker = self.pool.pop_idle()
+                return worker
+            await asyncio.sleep(1)
+
+    def push(self, worker):
+        """Push a worker back to the pool."""
+        self.pool.push(worker)
